+++ src/app/home/styles.css
@@ -14,7 +14,7 @@
 .canvas-outer {
   position: absolute;
   width: calc(var(--canvas-width) * 1%);
-  max-width: calc(85vh - (4 * var(--unit)));
+  max-width: calc(85vh - (4 * var(--space-4)));
   aspect-ratio: 1;
   @media screen and (max-width: 576px) {
     width: 95%;
